requirements.txt
streamlit
requests
pandas
python-dotenv
selenium
webdriver_manager
